Float scale factor

+++ kth/dim_14/cov_matrices/cov-mat-impl.hpp
@@ -1,7 +1,7 @@
 inline
 cov_mat_kth::cov_mat_kth( const std::string in_path,
 			  const std::string in_actionNames,  
-			  const int in_scale_factor, 
+			  const float in_scale_factor, 
 			  const int in_shift,
 			  const int in_scene, //only for kth
 			  const int in_segment_length
